Rebases 511 branch with current master and adds client fps memes.

+++ code/game/machinery/camera/camera.dm
@@ -105,6 +105,14 @@
 					O << "The screen bursts into static."
 			..()
 
+/obj/machinery/camera/tesla_act(var/power)//EMP proof upgrade also makes it tesla immune
+	if(isEmpProof())
+		return
+	..()
+
+/obj/machinery/camera/blob_act(obj/effect/blob/B)
+	if(B && B.loc == loc)
+		take_damage(health, BRUTE, 0)
 
 /obj/machinery/camera/ex_act(severity, target)
 	if(src.invuln)
@@ -162,6 +170,8 @@
 
 		else if(istype(W, /obj/item/device/analyzer))
 			if(!isXRay())
+				if(!user.drop_item(W))
+					return
 				upgradeXRay()
 				qdel(W)
 				user << "[msg]"
